+++ src/main/java/com/transcendruins/resources/sounds/Sound.java
@@ -99,12 +99,11 @@ public final class Sound {
      * Retrieves a sound from the available paths of this <code>Sound</code>
      * instance.
      * 
-     * @param random <code>double</code>: The random ID key to use, in the range of
-     *               <code>[0.0, 1.0]</code>.
+     * @param random <code>long</code>: The random ID key to use.
      * @param paths  <code>Map&lt;String, TracedPath&gt;</code>: The supplied paths.
      * @return <code>StoredSound</code>: The retrieved sound.
      */
-    public StoredSound getSound(double random, Map<String, TracedPath> paths) {
+    public StoredSound getSound(long random, Map<String, TracedPath> paths) {
 
         String path = entries.get(random);
